requirements.txt
snowflake-connector-python>=2.7.0
snowflake-sqlalchemy>=1.4.0
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
cryptography>=3.4.0
scikit-learn>=1.0.0  # Optional, for data science workflows 